legg tester på riktig sted
